+++ src/cmd/install.go
@@ -4,12 +4,13 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var installCommand = &cobra.Command{
-	Use:   "install",
-	Short: "Commands to set up Git Town on your computer",
-	Args:  cobra.NoArgs,
-}
-
-func init() {
-	RootCmd.AddCommand(installCommand)
+func installCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "install",
+		Short: "Commands to set up Git Town on your computer",
+		Args:  cobra.NoArgs,
+	}
+	cmd.AddCommand(aliasCommand())
+	cmd.AddCommand(completionsCmd())
+	return cmd
 }
